Store PDepNetwork path reactions with source/isomers as reactants.
We must do this because we are trying to minimize the density of states
calculations. The density of states of the reactants are needed to
perform the ILT calculation to get the microcanonical rates k(E). Since
these are only available for isomers and reactants, we must store the
path reactions in the proper direction in order for this to work. Note
that this will often require fitting reverse kinetics rather than
directly using those generated from the database.