Added log messages to solver for when rates exceed critical values.
The log messages briefly state which rate is the cause, as well as
printing the characteristic rate and the maximum species rate and
network leak rate, along with their fractions of the characteristic
rate. This makes it very easy to check that things are proceeding
smoothly. We also print the current value of any termination target
conversions to give a rough idea of how close to the "end" of the
simluation we are.