Better procedure for exploring an isomer of a partial network.
The previous version of addReactionToUnimolecularNetworks() was not
guaranteed to add reactions generated via exploring a well of a network
to that network. Instead, we explicity tell that function what network
to add the reaction to. This also saves some unnecessary computations.
If the network is not specified, the previous algorithm (which works for
when species are added to the model core) is still used.