Now treating all PDepReaction (path reactions) as irreversible.
Each partial network is responsible for producing k(T,P) values from its
source to all other configurations in that network in the forward
direction only. Another partial network will provide the reverse
direction if it should exist (e.g. if the net reaction is in the model
core). By setting the reversibility to False, we avoid double-counting
of these reactions. (We sacrifice guaranteed thermodynamic consistency
as a result, but the k(T,P) values should still be close.)

More work is still needed to ensure the right reactions are ending up in
the core and edge, and that duplicates are being avoided.