More modification of the enlarge algorithm to handle pressure dependence.
The most important change made here is that, when a species is added to
the model core, it must be explored in every partial network it is
contained in. This is necessary in order to get both the forward and
reverse PDepReactions in the model core.

I've created methods PDepNetwork.exploreIsomer and
CoreEdgeReactionModel.processNewReactions because we need these code
segments in multiple places.

Finally, I've added a few other minor (but important!) bugfixes and
checks.