Added getModelSize() method for determining the core and edge sizes.
This means that every time we want to know this information (e.g. for
printing to the console), it will be computed the same way. In
particular, we cannot simply rely on the lengths of the core and edge
species and reaction arrays because we would double-count the
PDepReactions (which exist as separate irreversible reactions in the
reaction lists), so we count each of those as half a reaction.