Added initial files for combined documentation for all subprojects.
It makes quite a bit of sense to combine the documentation for all
RMG Py subprojects into one single Sphinx project. In particular, this
allows for a single style and theme and for easy cross-referencing. (The
downside is that rebuilding the entire documentation will take a bit of
time, but this doesn't happen very often, so it shouldn't be an issue.)