Added custom style and layout to new combined documentation.
The theme is based on MIT colors, but with less contrast in the
background colors (white and light gray). MIT's maroon is used in lots
of places for accenting, while black remains the primary color for
regular text. The primary inspiration I used was Wikipedia.

Note that I have moved the related bar inside the body rather than above
and below the entire width of the page.