Renamed chempy package to rmgpy.chem; moved modules and unit tests.
The primary motivation is to place all of the RMG Py subprojects under a
single superpackage (named rmgpy). This means that we are no longer
keeping the subprojects in distinct directories, but combining them into
a massive project.

The combined project will share a single setup.py script and Makefile.