Renamed measure package to rmgpy.measure.
Note that, since modules in rmgpy.chem and rmgpy.measure share the same
name, the --pyrex-c-in-temp flag may not be usable because it does not
preserve the package structure when generating the C (and HTML) files.
This is unfortunate because I'd really rather not have intermediate
build files polluting my source directories.