Added first version of rmgpy.statesfit._fit module.
This module is intended to replace all of the Fortran functionality by
using my new PyDQED module to interact with the DQED optimization code.
The goal with this effort is to remove all non-Cythonized modules from
RMG Py; the statesfit package is the only one left.

The direct case seems to be working as well (or better!) than before. I
haven't tested the pseudo cases yet.