Added several unit tests for the new statesfit algorithm.
These unit tests check to see how well the fitting reproduces known data
for many combinations of vibrations and hindered rotors. All of the unit
tests pass. The conclusion seems to be that the algorithm does quite
well when it can fit the values directly, but less well when using
pseudo-vibrations and/or pseudo-rotors. This is what makes the group
frequency approach useful.