Significant rewrite of the rmgpy.data.statesfit module.
The motivation was to combine the setup and postprocess functions for
each of the three cases into single functions. Since the optimization
call was streamlined, this makes much more sense and results in
easier-to-follow code.

The other major change is that fitStatesToHeatCapacity() now returns a
list of HarmonicOscillator and/or HinderedRotor objects, rather than
lists of the vibrational frequencies and hindered rotor
frequency/barrier pairs.