Added first version of pruning algorithm.
This algorithm prunes all species below the flux tolerance to keep in
edge, and also prunes low-flux species until the maximum edge size is
achieved. The maximum maximum edge species rate across all reaction
systems is used for this evaluation. Pruning is only allowed when all
reaction systems are simulated to a valid termination condition, which
is why there is some value in simulating past the point where a model
enlarge event is identified as necessary. A few bugfixes in
removeSpeciesFromEdge() were needed to make that algorithm work.

Note that we still have some work to do to modify this algorithm for
handling pressure-dependence. We should also be wary of cases where zero
edge flux is observed.