Fixed bug in reaction model enlarging with pressure dependence.
When exploring isomers of partial networks to ensure that all isomers
that are core species are explored in every network, we were forgetting
to add the new reactions and species from this exploration to the
appropriate lists. This only matters when pruning is on, as otherwise
this process won't reveal any new reactions or species. But when pruning
is on, this process can regenerate new versions of pruned species and
reactions, so we need to do this to be safe.

Also added a slightly more informative check for this error condition
as each network is updated.