Now only pruning edge species that have existed for several iterations.
This hopes to minimize the effects of our one-species-at-a-time approach
to enlarging the model core. For a fast edge reaction A + B -> C + D, we
will likely add C and D in consecutive iterations, but until we add
both, this reaction does not exist in the core, so negligible C is made.
We don't want to prune products made from reacting C until D is also in
the core.

The minimum age for pruning eligibility is currently (and somewhat
arbitrarily) set to three. This may not capture every instance of the
above.