Fixed bug in computing leak rate for networks with one path reaction.
For these networks, we want to use the high-pressure-limit k(T) as the
leak rate to ensure that we're overestimating this value. However, the
leak direction need not be the direction in which the kinetics are
defined. This bugfix better handles both cases, and adds some error
checking in case neither direction has kinetics (which is clearly a
problem!) at leak flux evaluation time.