Refactored addSeedMechanismToCore() function.
The previous version had a couple of curious choices of variables to use
in a query, leading to poor readability and possibility of taking
improper action. For example, the previous algorithm was not properly
sorting the reactants and products of the forward and reverse reactions,
which explains the bug in checkForExistingReaction() addressed in a
recent prior commit. The new version is easier to follow and addresses
this and a few other inconsistencies.