Better handling of reaction libraries when pressure dependence is on.
We want to preserve the entirety of the seed mechanism as is when
placing it in the model core, including its pressure-dependent kinetics.
Thus, we cannot send these reactions to MEASURE.

For reaction libraries (and kinetics libraries), we want to use MEASURE
to compute the pressure-dependent kinetics, so we discard information
about the high-pressure-limit kinetics when loading these libraries.

Seed mechanisms + pressure dependence seems to be in working order as a
result of this commit.