Now converting colliders to Species objects when seed mechanism is loaded.
This is so that we can potentially use the collider efficiencies within
RMG simulations and successfully write them to output files (e.g.
Chemkin). Note that we haven't implemented collider efficiencies in the
solver yet, as there are some nontrivialities regarding this. In
particular, the way RMG Java handles colliders is not very flexible, and
we need to decide if we want to do it a better way.