Added __reduce__ method to Element class.
"Pickling" in Python is used to serialize and de-serialize Python
objects. This is a great way to save the current state of a Python
object in a way that we can easily reconstruct that state in the future.
Pickling should make implementation of restart functionality in RMG
almost trivially easy.

However, the pickling algorithm needs some help to know how to pickle
and unpickle Python extension types, e.g. cdef classes in Cython
modules. The __reduce__ method provides this help. This commit shows a
typical __reduce__ method, implemented for the (relatively simple)
Element class.