Added pickling/unpickling helper methods to rmgpy.chem extension types.
These methods are necessary for pickling and unpickling to be successful
in compiled Cython mode. (For the most part, pickling already worked in
pure Python mode, but RMG probably will never be run in a production
environment without compiling the chem modules.)

Reasonably extensive unit tests (coming in a separate commit) have
established that pickling of these extension types is successful in
compiled Cython mode.