Reenabled restart file reading and writing in RMG.
If saveRestart is True in the options block of the input file, RMG will
pickle the reaction model to a file after each set of enlarge actions
(i.e. before each simulation step). Saving the restart file is pretty
trivial now that we have pickling methods in place.

The hard part remains loading a restart file from a prior job. There are
several instances in the unpickled reaction model that point to the
species or database loaded from the last job, and need to be pointed to
the equivalent objects in the new job. In particular, species present in
the input file and kinetics databases are at fault. Some ugly code to
handle this has been written.