Bugfixes to the processQuantity() method to handle more varied input.
I was assuming that a list or tuple of length 2 was always a (value,
units) pair, when it could also be a (value1, value2) pair. A simple
test of units/value2 for string type allows proper handling of both
cases.

The processQuantity() method automatically converts lists to numpy
arrays, but a bug prevented it from converting matrices. This has been
addressed.

Added exception to processQuantity() for handling frequencies
passed in cm^-1. The quantities package wants to convert this to m^-1,
but this exception suppresses that conversion.

Finally, added a few units declarations to the rmgpy.chem.constants
module that are common in chemical kinetics problems but oddly not
defined in the quantities package as of today.