Improved __init__ and __repr__ methods in several rmgpy.chem classes.
By definition, the __repr__() method should return a string
representation that can be used to reconstruct an identical copy of the
object. This had not yet been done to many of the classes in the
rmgpy.chem.* modules.

Furthermore, we also want the __init__() method to be able to process
parameters with accompanying units, so that we can use it to process
input files in Python-formatted syntax using the exec command. This is
nice because we can now have the __repr__() method return a string with
units accompanying the various values, and still be able to process it
with the __init__() command. Thus we should be able to use __repr__() to
save output files.