Added unit tests for new rmgpy.chem __init__ and __repr__ methods.
These unit tests verify that there is negligible loss of data when
making a new version of an object based on its repr() output. This is
good, as it would be a shame to have gone through an extensive job (e.g.
RMG) correctly, only to save output that is incorrect or incomplete.