Use formula instead of SMILES for labels of RMG-generated species.
SMILES strings for complicated or large molecules can be quite long.
Furthermore, they can contain characters that may not interact well with
various outputs (e.g. the hash character in HTML). Thus it is better to
use the molecular formula, which is guaranteed to only consist of
alphanumeric characters and not be too long. We lose the ability to
distinguish between the various isomers except by their index, but as
long as we have pretty output this shouldn't be too much of a loss.