Added first version of HTML output generation for RMG output.
We want RMG to output information in a format that can be easily
digested by humans; web pages are a good choice for this because
everyone has a web browser. Richard wrote a tool for this purpose for
RMG-Java, but it is dependent on a very specific, customized set of
packages, making it tough to distribute. I am integrating his tool into
RMG Py such that it can be used by anyone, and so that it runs
automatically as part of the RMG program.

Currently the HTML output is simply a list of the reactions, with
drawing of the reactants and products using the
rmgpy.chem.ext.molecule_draw module. Much more to come!