Added javascript to output HTML show/hide reactions by family.
This gives users a means to view only a subset of the reaction model at
a time, which may make it easier to find individual reactions of
interest. The jQuery library is utilized for this purpose. Note that I'm
using check boxes instead of span blocks so that the user has a visual
cue which families are shown. I've also added check all and uncheck all
actions.