Fixed bug in fitting of frequencies to heat capacity data.
The bug was that the group frequencies were not being included in the
returned set of frequencies; only the fitted frequencies and rotors were
returned. Naturally, this left us way short of the required number of
modes!

Also fixed error in HinderedRotor.__repr__() which caused a segfault in
compiled Cython mode.