MEASURE input files saved during RMG runs once again run standalone.
Recently I changed the default label for species to be their formula
instead of their SMILES. This means that many species have the same
label, so you must also use the index in order to identify unique
isomers. This is particularly important in PDepNetworks, which often
have many isomers in common. Thus, when saving MEASURE input files, all
species are identified by their label *and* index.