Improved procedure for automatic determination of energy grains.
First we determine the highest ground-state energy (of a species or
transition state) on the potential energy surface. Next, for the
highest-ground-state-energy isomer, we compute the density of states and
determine the energy at which the Boltzmann distribution has tailed off
and can be truncated. Adding this to the highest ground-state energy
gives the desired Emax, which is now guaranteed to be above all energies
on the PES.