Now computing reactant channel density of states during RMG jobs.
Now that we are using the partial network/activated species algorithm
paradigm of Matheu et al, there is not much speed penalty for computing
the density of states for the reactant channel, as each partial network
has either 0 or 1 such channel. The k(E) calculation via ILT is also
more accurate when the density of states for the reactant channels is
available, so it appears to be worth the time. Finally, this ensures
that the standalone MEASURE jobs run in the same way as the jobs that
run during RMG.

Also cleaned up a log message and a breakpoint involving network
merging.