Now using pmutil instead of guppy for memory profiling.
My experience with heapy (the component of guppy used for memory
profiling) was that it was more featured but also slow and finicky. The
pmutil package seems to be much faster and to provide the level of
functionality we require, so I've switched to that. Note that pmutil is
optional, with the only drawback of not installing it being that you
won't get memory profile information.

Also cleaned up log messages for the execution time (now given in
HH:MM:SS) and restart file size, and promoted these to be printed at the
default level of verbosity.