Added first version of Quantity class.
The Quantity class is for storing numerical quantities with optional
units and uncertainties. The key here is to preserve the ability to work
with the quantities quickly; to do this, we need to store the values
internally in SI units so that we don't have to do unit conversions at
every computation. However, we store the original units so we can
convert back in certain circumstances (e.g. printing to file). This
class can also handle sets of values.

Also added some unit tests for the new Quantity class.