Fixes to enable successful Cythonization after recent changes.
Most notable is that we cannot use __eq__ in Cython mode, so no equality
testing is allowed between Quantity objects. Also, __reduce__ and
__setstate__ methods were added to the Quantity class to enable
successful pickling. It seems that sometimes a Quantity object is passed
to the Quantity __init__ method; I've added code to handle this by
simply making a copy of the object.