Another round of bugfixes regarding use of the new Quantity object.
Notable in this round is the restoration of MEASURE to basic working
functionality. The Quantity object makes the input file parsing more
streamlined, which is nice.

There is probably still quite a bit of functionality that needs
processing to work with Quantity objects, but we're slowly getting
there. I expect at least a few of these things to be nontrivial to
adapt, for various reasons. My hope is that direct use of Quantity
objects remains as behind-the-scenes as possible.