Yet another round of modifications involving use of Quantity objects.
The primary impetus in this round was to get RMG Py running again, which
seems to have been achieved. In the future we will need to actually
think in more detail about how exactly to best use Quantity objects, but
for now, I'm happy with most things working more or less as before. RMG
doesn't seem to be too much slower than before, which is important.