More updates and bugfixes to the documentation.
These fixes should squash just about all of the warnings that Sphinx
currently raises. Most of those referred to classes and functions that
no longer exist due to recent changes.

Note that you must compile MEASURE before you compile the documentation
for the MEASURE API reference to be fully successful (as several of the
modules are only Cython modules). The same will need to be done for the
solver once someone gets around to writing that documentation.

The MEASURE theory guide takes quite a while to build due to all of the
TeX equations. Not sure if jsMath would be better for this.