Added first version of new Entry class.
The new Entry class should be able to represent any entry in any
component of the RMG database. The actual content of the data members
(particularly the item and data attributes) will vary depending on which
component of the database you are looking at. For instance, item could
be a species, functional group, or a reaction, and data could be a
thermo or kinetics object.

The Entry class replaces the DataEntry class.