Added rmgpy.data.reference module for representing bibliographic information.
The classes contained herein - Reference, Article, Book, and Thesis -
are based on BiBTeX format. These classes should represent the majority
of the reference types we will encounter in the database, but we can
always add more as needed.

Currently each class has a __str__ method that provides a nice,
reStructuredText-formatted string representation of the reference
information. I am not too concerned about the bibliographic format at
this time; as long as the user provides a URL (e.g. a DOI), then the
bibliographic format is less important.

Adding URLs is strongly encouraged to make it very easy to look up the
original reference. This is important for documenting the data, so that
it is easier to verify in the future.