Merged Dictionary, Tree, and Library classes into Database class.
The Entry class stores all of the information that each of these
components refers to, so it makes more sense to think of things as a
single database than as a set of components. The load and save methods
for both the new and old database formats have been preserved, albeit
with some modifications.

The Database class now serves as a base class. There are several methods
that you must implement in the derived class to be able to load and save
the database.