Added new framework for thermo database.
The thermo database consists of a depository, a set of libraries, and a
set of group additivity values. Each of these has its own separate class
that derives from the Database base class. To reduce code duplication, a
few module-level functions have been created that the corresponding
methods of these classes call. There is also a master ThermoDatabase
class that automatically handles the depository, libraries, and groups
as a contiguous set.

Methods have been written to enable loading and saving of the thermo
database in both the old (RMG-Java) and new formats. These seem to be in
working order, although they may not yet be feature complete. The new
database format is probably still going to change a bit in upcoming
commits; thus it is important to preserve being able to read and write
to both the old and new formats.