Added new framework for kinetics database.
The kinetics database consists of a depository, a set of libraries, and
a set of group additivity values. Each of these has its own separate
class that derives from the Database base class. To reduce code
duplication, a few module-level functions have been created that the
corresponding methods of these classes call. There is also a master
KineticsDatabase class that automatically handles the depository,
libraries, and groups as a contiguous set.

Methods have been written to enable loading and saving of the kinetics
database in the new format and loading from the old format.