Enabled reading old kinetics groups and reading/writing new kinetics groups.
This is the first step in merging the ReactionFamily class with the
KineticsGroups class. Note that reaction families have additional
metadata to be handled, so I've overriden the load() and save() methods
of the Database base class. This causes a bit of code duplication, but
is necessary.

Note that saving in the old format is not currently supported, as it
involves some extra work. Also note that we do not keep the data from
the kinetics groups rate rules, as we will be replacing them with group
additivity values in the very near future.