Added KineticsData class to rmgpy.chem.kinetics module.
The KineticsData class is used to represent rate coefficient data
consisting of a discrete set of rates k(T) at various temperatures. This
is analogous to the ThermoData class, which does the same for heat
capacities. The idea is to preserve the ability to manipulate and fit
the data for future use, and to provide an interpolation model for cases
when other models don't provide a good fit.

Note that KineticsData currently stores high-pressure limit k(T) values
only.