Better handling of metadata when loading database.
In particular, the metadata that is set when you load the new-style
format is *not* set when loading the old database. This is unfortunate
for when we are converting from the old format to the new format. This
commit provides default values for the label and name of each database
component so that these attributes are always set in the converted
database.

Some updates to the constructors of the database components enable easy
setting of these attributes at construction time.