Better handling of special collider species when loading old database.
The species N2, He, Ne, and Ar are treated as special "bath gas" species
in the old RMG database, and are therefore not required to be explicitly
defined in the species dictionary of the kinetics library. The new
database does not make a distinction for these special bath gas species,
so this commit adds a bit of hardcoded functionality to handle the
special colliders. The warning for unknown collider remains in case any
future hardcoding of bath gas species in the old RMG database occurs
(unlikely).