Improvements to the loadOldLibrary() functionality.
The code for processing the library file was moved to a new method
parseOldLibrary(). This was done because, when loading the old kinetics
database, we are interested in loading the old library but not in
storing those values in the regular entries. (In fact, as the database
is currently designed, we could not do so even if we wanted to.) The
original loadOldLibrary() method works as before, and contains a call to
parseOldLibrary().

Better handling of the (optional) index column of the library files and
the columns of parameters was also added.