Added functionality for fitting kinetics group values from old rate rules.
The main function of interest is the fitGroupValuesFromOldLibrary()
method of KineticsGroups. This function wraps calls to parseOldLibrary()
for loading the old library file and fitGroupValues() for performing the
group value fitting. Lots of data is returned by this method so that we
can use it to make tables and figures.

This requires parsing the old kinetics rateLibrary.txt file containing
the rate rules, which had been broken for some time because we can't
really use the old rate rules directly in the new database. The
processOldLibraryEntry() method takes care of most of this. Note the
unfortunate bit of hardcoding of reaction families required to get the
units of the rate coefficients correct.