Restored functionality for generating thermo data from database.
Thermo data can be generated from each of the three sections of the
thermo database (depository, libraries, and groups), either by finding
the first match or by finding all matches. A significant fraction of
this code is restored from earlier work that was inadvertently deleted
while in the process of refactoring the database code on this branch.