Restored functionality for generating reactions from the kinetics database.
This includes both finding reactions in the depository and libraries,
and generating reactions from the templates in the reaction families.
Like the previous commit, much of this is simply restoring code that was
deleted in order to make refactoring the data package easier for me to
do.

Some work toward generating kinetics for the reactions is also present,
but this is not really in working order due to the inability to generate
kinetics from group values (since we don't yet have any group
additivity values to use).