Import needed packages before exec command during database loading.
This is so that we can load the database in compiled Cython mode, which
doesn't allow package imports that occur as a result of using the exec
command. In particular, this was noticed when reading ThirdBody and
derived classes, as colliders are represented as SMILES strings which
require OpenBabel to process.