Now silently dropping generated reactions that have wrong number of products.
In order to fully match the template, the generated reactions must also
produce exactly the number of products as indicated by the template.
This is required so that we can distinguish between intermolecular and
intramolecular versions of a reaction family (e.g. R_Recombination and
Birad_recombination). There are several reaction families that
potentially fall victim to this (including a few that don't have an
intramolecular equivalent currently available in RMG).

Unfortunately, by silently dropping these reactions, we may be squashing
errors due to actual bugs in the reaction templates.