Added script for importing RMG-Java database to the new RMG-Py format.
Most of the action for this script happens in the various loadOld()
methods of the thermo and kinetics database classes. Nonetheless, this
script is useful as a convenient place to do the conversion from the old
to the new format.

Note that we don't quite have a complete conversion yet, as we have no
way of determining kinetics using rate rules or group additivity values
yet.