Now fitting kinetics group values when importing old RMG database.
We are able to fit at least some group values for every family from the
rate rules, although a few families are so sparse that the uncertainties
are very large. The process for fitting group values can be rather slow
due to the expensive fitting procedure. The fitted group values are
stored in the kinetics group database and can be saved to file and
loaded from file successfully.