Removed getKinetics() methods from KineticsDatabase class.
You would think that such methods would be desirable, as they appear to
parallel the getThermo() methods of the thermo database. However, it
turns out that this isn't a very well-defined means of interacting
with the kinetics database, as you really need reaction families and
labeled central atoms to make it work, which you generally don't have.
Instead, the parallel is much more closely to the generateReactions()
methods, which generate the kinetics along the way.