Added options for loading libraries and depository to database classes.
These give the user finer control over what components of the database
are loaded. In particular, the user can choose to only load the
components necessary for the desired task. In RMG jobs, for example, the
user doesn't need to load the depositories, and probably only needs
certain libraries.

The order of specification of the libraries is now respected when
getting thermo data and generating reactions from the database.