Updated database section of RMG input files.
To specify the database, you now specify its root directory, as well as
the thermo and kinetics libraries to load. The latter is split into seed
mechanisms and reaction libraries based on where the loading reactions
are placed. You don't need to specify the thermo and kinetics groups any
longer; these are always loaded.

The example input files have been updated to reflect the new syntax.