Refactored main RMG code to use new database classes.
The presence of a singular RMGDatabase class means that we don't need a
bunch of global objects to represent the individual database components.
The individual ThermoDatabase and KineticsDatabase classes mean that we
don't need module-level methods either.

The new KineticsDatabase.generateReactions() method returns reactions
containing Molecule objects. (Before we were expecting Species objects
to be returned.) This means that we needed to add code to create Species
objects from the Molecule objects and to check for reaction uniqueness.
Attributes newReactionList and newSpeciesList have been added to
CoreEdgeReactionModel to store the new species and reactions since the
last enlarge().

Note that we are not storing the database as an attribute of
CoreEdgeReactionModel because we don't want to include the database in
the restart file.