Added getSource() to DepositoryReaction, LibraryReaction, and TemplateReaction.
This method returns the database component that was the source of the
reaction. Since each object calls the attribute that contains this
information by a different name, we need a common method to retrieve
this information. (This is a much better approach than doing
isinstance() checks, as it leverages the object-oriented approach more
cleanly.)