Fixed bug in KineticsGroups.getKinetics() method.
When adding in the more specific corrections from the actual fitted
functional groups, we need to check that we don't add in corrections
from top-level nodes, as these are by definition zero. (We store the
generic kinetics in these nodes, so we can't just set them to zero.) The
place to compare with should be the top attribute, not the
forwardTemplate.reactants attribute.