Added new framework for RMG states (frequencies) database.
This database provides means for determining the molecular degrees of
freedom for a given molecule, either via lookup in a library or
estimation by fitting to heat capacity data. The structure looks very
similar to the other database components, particularly the thermo
database.

Currently the old frequency database can be loaded and the new frequency
database loaded and saved.