Refactoring to (nearly) restore pressure dependence functionality in RMG.
The code that generates the molecular degrees of freedom and the code
that handles placing path reactions in networks needed some refactoring
in order to work with the new database. For the most part this seems to
be working. However, there is a feature missing from MEASURE to handle
path reactions of the form productA + product B <- isomer; that is, the
kinetics are defined in the association direction, but we want to treat
the reaction as an irreversible dissociation. This will take some
additional effort to decide how to handle.