Moved rmgpy.chem.constants module to rmgpy.quantity.
The constants module doesn't really have anything to do with "chem", so
it doesn't belong in that subpackage. (This was left over from when
there was a separate project ChemPy for these modules.) I've also
renamed it to quantity to reflect that it contains more than just
constants, i.e. the Quantity class.

Also moved the corresponding unit test file accordingly.