Collected physical constants into constants object.
The idea here is to enable importing of the constants as a whole via the
statement ::

from rmgpy.quantity import constants

so that code that uses, e.g., constants.R will not need to be changed.
To do this, I have created a Constants class to collect the physical
constants, and declared a module-level variable called constants that is
an instance of a Constants object. Note that this also allows
declaration of the attributes in Cython, although I don't anticipate
any Cython acceleration as a result, since constants is still a Python
object.