Significant refactoring of the Quantity class.
Most of the refactoring effort was focused on the __init__() method. The
goal here was to increase the constructor's functionality so it could
handle quantities packed in tuples, quantities unpacked from tuples,
Quantity objects, and simple numbers.

The getConversionFactorToSI() and getConversionFactorFromSI() functions
are now methods of the Quantity class, instead of module-level
functions.

Lots of documentation has been added and rewritten as well.