Refactoring and cleanup of the various thermodynamics model classes.
In the Wilhoit class, the polynomial coefficients a0, a1, a2, and a3 are
now of type double instead of Quantity because they are dimensionless
and have no uncertainties associated with them. Mostly this is a
philosophical change, as the readability and speed of the code is
largely unaffected by this choice.

String formatting now uses the syntax recommended for Python 3, as the
syntax using the "%" character is considered deprecated. The new
formatting syntax is a little more verbose, but a lot more powerful. If
we ever want to run RMG Py in Python 3, this will be very important.

Lots of updates to the docstrings and other documentation as well.