Moved relevant functions from thermo_converter.py to thermo.py.
The thermo conversion functions are closely linked to the thermo
classes, so they belong together in the same module. I had separated
them because the former are quite ugly (and necessarily so, not the
fault of the programmer!), but I don't think this is a valid reason to
keep them apart. Note that only the methods that we actually use are
kept; the others will be deleted (but remain in the history so we can
recover them if we change our mind).