Added unit tests for the various thermo model conversions.
Conversion between thermo models inherently introduces additional
uncertainty due to the approximate nature of the mapping of the new
model parameters to the old model data. Thus, the unit tests only check
for order-of-magnitude correctness in the new and old model's
thermodynamic quantities. This is okay for now, but it would be better
if we could make these tests more strict in the future.