Added and refactored lots of unit tests for the rmgpy.kinetics module.
With this commit every method of each kinetics model class has a
corresponding unit test, and all of them pass successfully. Along the
way these unit tests helped me find and address a few of the bugs fixed
in the last commit, including a few that would have been less than
obvious to spot in other settings. The value of test-driven development!