Extensive refactoring of the rmgpy.statmech module.
As usual, this is to clean up and update documentation, find and fix
bugs, and update to Python 3 string formatting syntax. A few subtle bugs
were identified and fixed. The functionality and use should be largely
unchanged otherwise.

The statmech module also had not yet been converted to use the Quantity
class, so I took care of that as well. Not sure why this hadn't been
done yet (presumably laziness on the part of this developer).